+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added at [0.2.1]
 
+- **Recursive Deletion:** Implemented recursive deletion for narrative items.
 - **File Attachment:** Added the ability to attach auxiliary files to a world object and access them from the UI.
 - **Auto-save:** Added auto-save functionality when switching between items.
 - **Narrative Plan:** Implemented mass addition of narrative plan items from pasted text.